requirements.txt
pydantic==2.7.2
fastapi==0.111.0
fastapi-cli==0.0.4
psycopg2==2.9.9
pydantic==2.7.2
python-dotenv==1.0.1
